+++ misskey-api/src/model/note_favorite.rs
@@ -16,3 +16,5 @@ pub struct NoteFavorite {
     pub note_id: NoteId,
     pub note: Note,
 }
+
+impl_entity!(NoteFavorite, NoteFavoriteId);