+++ misskey-api/src/model/user_list.rs
@@ -18,3 +18,5 @@ pub struct UserList {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub user_ids: Option<Vec<UserId>>,
 }
+
+impl_entity!(UserList, UserListId);